+++ src/scripts/index.js
@@ -1,8 +1,6 @@
-import _ from 'lodash';
+// import _ from 'lodash';
 import '../assets/styles/base.scss'
 import './keyboard'
 
 import './description';
-
-
